backend/requirements.txt
fastapi 
pydantic
uvicorn[standard] 
sqlalchemy 
psycopg2-binary

jinja2

python-multipart
python-dotenv
python-jose[cryptography]

passlib
bcrypt==4.0.1

alembic

email-validator
